# gateway/openapi-spec.yaml
openapi: 3.0.3
info:
  title: ERP AI Delay Risk API
  description: |
    Enterprise API for supply-chain delay risk scoring, feature retrieval,
    and high-risk notifications. Governed by Apigee API Gateway.

    **Authentication:** API key required in `X-API-Key` header.
    **Rate Limit:** 100 requests/minute per API key (standard tier).
    **Versioning:** URI-based — all routes prefixed with `/v1/`.

    **Architect:** Dr. Tatianna Gilliam — Cloud & AI Architect (AZ-305 | AI-102 | AZ-104)
  version: 1.0.0
  contact:
    name: API Platform Team
    email: api-platform@example.com
  license:
    name: Proprietary
servers:
  - url: https://api.delayrisk.example.com/v1
    description: Production (Apigee)
  - url: https://api-staging.delayrisk.example.com/v1
    description: Staging
security:
  - ApiKeyAuth: []
...
paths:
...
  /features/{entity_id}:
    get:
      tags: [Features]
      summary: Retrieve features for an entity
      description: |
        Returns the online feature vector for a single entity. Backed by
        Vertex AI Feature Store in production; local cache in portfolio demo.
      operationId: getFeatures
      parameters:
        - $ref: '#/components/parameters/EntityId'
        - $ref: '#/components/parameters/RequestIdHeader'
      responses:
        '200':
          description: Feature vector retrieved
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureResponse'
              example:
                entity_id: PO-88421
                features:
                  supplier_lead_time_days: 45
                  historical_delay_rate: 0.32
                  inventory_coverage_days: 12
                  order_value_usd: 125000
                  seasonality_index: 1.4
                last_updated: '2026-06-08T12:00:00Z'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /features/batch:
    post:
      tags: [Features]
      summary: Bulk feature retrieval
      description: Retrieve feature vectors for multiple entities in one request.
      operationId: getFeaturesBatch
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchFeatureRequest'
            example:
              entity_ids:
                - PO-88421
                - PO-88422
                - PO-88423
      responses:
        '200':
          description: Feature vectors retrieved
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchFeatureResponse'
              example:
                features:
                  - entity_id: PO-88421
                    features:
                      supplier_lead_time_days: 45
                      historical_delay_rate: 0.32
                    last_updated: '2026-06-08T12:00:00Z'
                  - entity_id: PO-88422
                    features:
                      supplier_lead_time_days: 21
                      historical_delay_rate: 0.08
                    last_updated: '2026-06-08T12:00:00Z'
                request_id: req_feat_batch01
                timestamp: '2026-06-08T14:32:00Z'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key issued via the Apigee Developer Portal.
        Standard tier: 100 req/min. Enterprise tier: 1000 req/min.
  parameters:
    EntityId:
      name: entity_id
      in: path
      required: true
      description: Unique entity identifier (purchase order, shipment, etc.)
      schema:
        type: string
        example: PO-88421
    RequestIdHeader:
      name: X-Request-ID
      in: header
      required: false
      description: Client-supplied correlation ID for distributed tracing
      schema:
        type: string
        example: req_a1b2c3d4e5f6
  headers:
    X-Request-ID:
      description: Correlation ID for the request
      schema:
        type: string
        example: req_a1b2c3d4e5f6
    X-RateLimit-Remaining:
      description: Remaining requests in the current rate limit window
      schema:
        type: integer
        example: 87
...
  schemas:
...
    FeatureResponse:
      type: object
      required: [entity_id, features, last_updated]
      properties:
        entity_id:
          type: string
        features:
          type: object
          additionalProperties: true
        last_updated:
          type: string
          format: date-time
    BatchFeatureRequest:
      type: object
      required: [entity_ids]
      properties:
        entity_ids:
          type: array
          minItems: 1
          maxItems: 500
          items:
            type: string
    BatchFeatureResponse:
      type: object
      required: [features, request_id, timestamp]
      properties:
        features:
          type: array
          items:
            $ref: '#/components/schemas/FeatureResponse'
        request_id:
          type: string
        timestamp:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required: [error_code, message, request_id, timestamp]
      properties:
        error_code:
          type: string
          description: Machine-readable error identifier
          example: RATE_LIMIT_EXCEEDED
        message:
          type: string
          description: Human-readable error description
          example: Rate limit exceeded. Retry after 42 seconds.
        request_id:
          type: string
          example: req_a1b2c3d4e5f6
        timestamp:
          type: string
          format: date-time
          example: '2026-06-08T14:32:00Z'
        details:
          type: object
          additionalProperties: true
          description: Optional structured error context
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: INVALID_REQUEST
            message: entity_id is required
            request_id: req_a1b2c3d4e5f6
            timestamp: '2026-06-08T14:32:00Z'
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: UNAUTHORIZED
            message: Invalid or missing API key
            request_id: req_a1b2c3d4e5f6
            timestamp: '2026-06-08T14:32:00Z'
    NotFound:
      description: Entity not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: ENTITY_NOT_FOUND
            message: No features found for entity PO-99999
            request_id: req_a1b2c3d4e5f6
            timestamp: '2026-06-08T14:32:00Z'
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          description: Seconds until the rate limit resets
          schema:
            type: integer
            example: 42
        X-RateLimit-Remaining:
          schema:
            type: integer
            example: 0
        X-RateLimit-Limit:
          schema:
            type: integer
            example: 100
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: RATE_LIMIT_EXCEEDED
            message: Rate limit exceeded. Standard tier allows 100 req/min.
            request_id: req_a1b2c3d4e5f6
            timestamp: '2026-06-08T14:32:00Z'
...